backend/requirements.txt
astor==0.8.1
Babel==2.9.1
bce-python-sdk==0.8.64
cachetools==5.0.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.10
click==8.0.3
cssselect==1.1.0
cssutils==2.3.0
cycler==0.11.0
Cython==0.29.26
decorator==5.1.0
distlib==0.3.4
et-xmlfile==1.1.0
fasttext==0.9.1
filelock==3.4.2
flake8==4.0.1
Flask==2.0.2
Flask-Babel==2.0.0
Flask-Cors==3.0.10
fonttools==4.28.5
future==0.18.2
gunicorn==20.1.0
huggingface-hub==0.2.1
identify==2.4.1
idna==3.3
imageio==2.13.5
imgaug==0.4.0
importlib-metadata==4.2.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
lmdb==1.3.0
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
networkx==2.6.3
nltk==3.6.7
nodeenv==1.6.0
numpy==1.21.5
opencv-contrib-python==4.4.0.46
opencv-python==4.5.5.62
openpyxl==3.0.9
packaging==21.3
paddleocr==2.3.0.2
paddlepaddle==2.2.1
pandas==1.3.5
Pillow==8.4.0
platformdirs==2.4.1
pre-commit==2.16.0
premailer==3.10.0
protobuf==3.19.1
pybind11==2.9.0
pyclipper==1.3.0.post2
pycodestyle==2.8.0
pycryptodome==3.12.0
pyflakes==2.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
regex==2021.11.10
requests==2.27.1
sacremoses==0.0.46
scikit-image==0.19.1
scipy==1.7.3
Shapely==1.8.0
shellcheck-py==0.8.0.3
six==1.16.0
textblob==0.17.1
tifffile==2021.11.2
tokenizers==0.10.3
toml==0.10.2
torch==1.10.1
tqdm==4.62.3
transformers==4.15.0
typing_extensions==4.0.1
urllib3==1.26.7
virtualenv==20.13.0
visualdl==2.2.2
Werkzeug==2.0.2
wordsegment==1.3.1
zipp==3.7.0
